+++ docs/en/CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to
 
 #### Breaking Changes
 #### Added
+- Add OpenSSL depends and add fmd5sum() function.
 - Add header file /usr/include/ulpatch/version.h.
 - rpmbuild.sh add --nocheck argument.
 - task: Add C macro `task_vdso_vma(task)`